add golden file test for tcnp resource (#1783)
