+++ service/controller/clusterapi/v29/resource/tcnp/template/template_main_iam_policies.go
@@ -1,7 +1,7 @@
 package template
 
 const TemplateMainIAMPolicies = `
-{{ define "iam_policies" }}
+{{- define "iam_policies" -}}
   NodePoolRole:
     Type: AWS::IAM::Role
     Properties:
@@ -13,7 +13,6 @@ const TemplateMainIAMPolicies = `
           Principal:
             Service: {{ .IAMPolicies.EC2ServiceDomain }}
           Action: "sts:AssumeRole"
-
   NodePoolRolePolicy:
     Type: "AWS::IAM::Policy"
     Properties:
@@ -26,35 +25,28 @@ const TemplateMainIAMPolicies = `
           - Effect: "Allow"
             Action: "ec2:Describe*"
             Resource: "*"
-
           - Effect: "Allow"
             Action: "ec2:AttachVolume"
             Resource: "*"
-
           - Effect: "Allow"
             Action: "ec2:DetachVolume"
             Resource: "*"
-
-          {{ if .IAMPolicies.KMSKeyARN }}
+          {{- if .IAMPolicies.KMSKeyARN -}}
           - Effect: "Allow"
             Action: "kms:Decrypt"
             Resource: "{{ .IAMPolicies.KMSKeyARN }}"
-          {{ end }}
-
+          {{- end -}}
           - Effect: "Allow"
             Action:
               - "s3:GetBucketLocation"
               - "s3:ListAllMyBuckets"
             Resource: "*"
-
           - Effect: "Allow"
             Action: "s3:ListBucket"
             Resource: "arn:{{ .IAMPolicies.RegionARN }}:s3:::{{ .IAMPolicies.S3Bucket }}"
-
           - Effect: "Allow"
             Action: "s3:GetObject"
             Resource: "arn:{{ .IAMPolicies.RegionARN }}:s3:::{{ .IAMPolicies.S3Bucket }}/*"
-
           - Effect: "Allow"
             Action:
               - "ecr:GetAuthorizationToken"
@@ -65,12 +57,11 @@ const TemplateMainIAMPolicies = `
               - "ecr:ListImages"
               - "ecr:BatchGetImage"
             Resource: "*"
-
   NodePoolInstanceProfile
     Type: "AWS::IAM::InstanceProfile"
     Properties:
       InstanceProfileName: gs-cluster-{{ .IAMPolicies.Cluster.ID }}-profile-{{ .IAMPolicies.NodePool.ID }}
       Roles:
         - Ref: "NodePoolRole"
-{{ end }}
+{{- end -}}
 `
