+++ service/controller/clusterapi/v29/resource/tcnp/template/template_main_subnets.go
@@ -1,8 +1,8 @@
 package template
 
 const TemplateMainSubnets = `
-{{ define "subnets" }}
-  {{ range .Subnets.List }}
+{{- define "subnets" -}}
+  {{- range .Subnets.List -}}
   {{ .Name }}:
     Type: AWS::EC2::Subnet
     Properties:
@@ -21,6 +21,6 @@ const TemplateMainSubnets = `
     Properties:
       RouteTableId: {{ .TCCP.Subnet.RouteTable.Name }}
       SubnetId: {{ .TCCP.Subnet.Name }}
-  {{ end }}
-{{ end }}
+  {{- end -}}
+{{- end -}}
 `
